Add function to generate truth table
Context: Up until now the utility only supports generating a function for the output of a truth table.

Problem: It is useful to generate a truth table to work with in figuring out the logic for a set of variables. This is not possible in the current state.

Fix: Update the utility to add a function which generates a truth table, also allowing the truth table to be exported to CSV for use in tools like sheets.